Added some rough draft templates for content
All of the content from the RPGs needs to be added to the tool, which so far I'm planning on creating a tool to do manually. It'll be time consuming, but the tool will also allow for users to homebrew content as well. These files are incredibly simple rough drafts to outline the kind of idea I'm going for.